requirements.txt
GitPython
pyyaml
py2neo